Add a local settings file to the 1st exercise
In this file I can store all the settings that are not useful in other
exercises, but only in the first one.

In this specific case, I need the Tikz package (LaTeX) to work with
some images of the 1st exercise. Moreover, I defined some new colors
that can be used for various tasks.