Add the proposition to be proved in 1st exercise
I added the text of the exercise (you can find it in the
proposition.tex file).

In order to do that, I have created a new theorem environment called
"Proposition". You can find it in the local-settings.tex file.

I have also added a command to represent a line segment. I added it in
a file called symbols.tex. In this file you should put all the macros
that define some symbols that are useful in this exercise.